+++ src/main/java/org/mmarini/wheelly/apps/RobotCheckUp.java
@@ -72,7 +72,7 @@ public class RobotCheckUp {
     public static final int ROTATION_TOLERANCE = 5;
     public static final double DISTANCE_TOLERANCE = 0.1;
     public static final Dimension RESULT_SIZE = new Dimension(800, 600);
-    public static final String CHECKUP_SCHEMA_YML = "https://mmarini.org/wheelly/checkup-schema-0.1";
+    public static final String CHECKUP_SCHEMA_YML = "https://mmarini.org/wheelly/checkup-schema-1.0";
     private static final Logger logger = LoggerFactory.getLogger(RobotCheckUp.class);
     private static final int TEST_SPEED = MAX_PPS / 2;
 
@@ -305,7 +305,7 @@ public class RobotCheckUp {
      */
     private void run() {
         logger.info("Robot check started.");
-        controller = Yaml.fromFile(parseArgs.getString("config"), CHECKUP_SCHEMA_YML);
+        controller = AppYaml.fromFile(parseArgs.getString("config"), CHECKUP_SCHEMA_YML);
         controller.readErrors().doOnNext(err -> {
             comMonitor.onError(err);
             logger.atError().setCause(err).log();